refactor(attribution): address module review feedback from PR #3316
- Move attribution.ts from src/hooks/ to src/utils/ (pure module, not a hook)
- Move signUpDataSchema to src/server/schemas/ (fix EE→tRPC layering violation)
- Add console.warn with dedupe for silent sessionStorage errors
- Add compile-time assertion ensuring zod schema covers all Attribution fields

Closes #454
