+++ langwatch/ee/billing/types.ts
@@ -2,7 +2,7 @@ import type { z } from "zod";
 import type { PlanInfo } from "../licensing/planInfo";
 import type { LimitType } from "../../src/server/license-enforcement/types";
 import type { PlanTypes } from "./planTypes";
-import type { signUpDataSchema } from "../../src/server/api/routers/onboarding/schemas/sign-up-data.schema";
+import type { signUpDataSchema } from "../../src/server/schemas/sign-up-data.schema";
 
 export type BillingPlanProvider = {
   getActivePlan(
