+++ langwatch/src/server/api/routers/onboarding/index.ts
@@ -1,2 +1,1 @@
 export * from "./onboarding.router";
-export * from "./schemas/sign-up-data.schema";
